api/requirements.txt
--index-url https://mirrors.cloud.tencent.com/pypi/simple

gunicorn
opencv-python==4.11.0.86
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.3.1
backoff==2.2.1
bce-python-sdk==0.9.29
beautifulsoup4==4.13.4
billiard==3.6.4.0
blinker==1.9.0
blis==1.2.1
cachetools==5.5.2
catalogue==2.0.10
celery==5.2.7
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
concurrent-log-handler==0.9.26
confection==0.1.5
contourpy==1.3.2
cos_python_sdk_v5==1.9.36
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dashscope==1.23.1
dataclasses-json==0.6.7
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
duckduckgo_search==8.0.1
email_validator==2.2.0
emoji==2.14.1
eval_type_backport==0.2.2
eventlet==0.34.1
executing==2.2.0
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flask_weaviate==1.1.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gevent==25.5.1
greenlet==3.2.2
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
injector==0.22.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kombu==5.5.3
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.65
langchain-deepseek==0.1.3
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-mcp-adapters==0.1.1
langchain-ollama==0.3.3
langchain-openai==0.3.23
langchain-postgres==0.0.14
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langchain-weaviate==0.0.4
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.63
langsmith==0.3.45
language_data==1.3.0
lark==1.2.2
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.9.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.18
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.8
onnx==1.18.0
onnxruntime==1.19.2
openai==1.75.0
optionaldict==0.1.2
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.3.0
parso==0.8.4
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20250506
pgvector==0.3.6
phonenumbers==8.13.55
pi_heif==0.22.0
pillow==11.0.0
pluggy==1.5.0
portalocker==3.1.1
preshed==3.0.9
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg==3.2.6
psycopg-pool==3.2.6
psycopg2-binary
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
pyreadline3==3.5.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pywin32==310;platform_system == "Windows"
PyYAML==6.0.2
qianfan==0.4.12.3
RapidFuzz==3.13.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tenacity==8.5.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tldextract==5.2.0
tokenizers==0.21.1
torch @ https://mirrors.aliyun.com/pytorch-wheels/cpu/torch-2.6.0+cpu-cp311-cp311-linux_x86_64.whl
torchaudio @ https://mirrors.aliyun.com/pytorch-wheels/cpu/torchaudio-2.6.0+cpu-cp311-cp311-linux_x86_64.whl
torchvision @ https://mirrors.aliyun.com/pytorch-wheels/cpu/torchvision-0.21.0+cpu-cp311-cp311-linux_x86_64.whl
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.2
typing-inspect==0.9.0 
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2

unstructured==0.17.2
unstructured-client==0.33.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.4.0
uvicorn==0.34.2
validators==0.34.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==4.13.2
webencodings==0.5.1
websocket-client==1.8.0
wechatpy==1.8.18
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
WTForms==3.2.1
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.19.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
